+++ src/main/kotlin/at/raven/ravenAddons/features/pit/CarePackageHighlighter.kt
@@ -1,4 +1,4 @@
-package at.raven.ravenAddons.features
+package at.raven.ravenAddons.features.pit
 
 import at.raven.ravenAddons.config.ravenAddonsConfig
 import at.raven.ravenAddons.event.render.container.ContainerBackgroundDrawEvent
